requirements.txt
streamlit
python-dotenv
requests
PyPDF2
python-docx